added token to api requests

+++ frontend/src/components/WorkoutDetails.js
@@ -2,15 +2,24 @@ import React from "react";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 import { FaTrash } from "react-icons/fa";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
-    if(window.confirm("Delete Workout?")){
-      const response = await fetch("/api/workouts/" + workout._id, {
+    if (!user) {
+      return;
+    }
+    if (window.confirm("Delete Workout?")) {
+      const response = await fetch("/api/workouts/" + workout.id, {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       });
+
       const json = await response.json();
 
       if (response.ok) {
